refactor: 方法toggleMovieModal在多處使用到，因此獨立出來放入mixins
